adddd

+++ ast/private.h
@@ -9,7 +9,7 @@ t_cmd	*create_cmd(int cmd_len, int redir_count, t_token *token);
 t_cmd	*consume_cmd(t_token **token);
 int 	consume(t_token_kind kind, t_token **token);
 
-int		ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 char	*ft_strdup(const char *s);
 
 #endif
